fixed a number of compiler warnings including new lines at the end of the file, out of order CIL, implicit casts, and unused variables.

+++ delta3d/src/dtActors/fenceactor.cpp
@@ -406,7 +406,7 @@ namespace dtActors
                         float fullLen = (start - end).length();
 
                         float index = posLen / mPostMaxDistance;
-                        subIndex = floor(index);
+                        subIndex = int(floor(index));
 
                         float len = fullLen - (subIndex * mPostMaxDistance);
                         if (len <= mPostMinDistance)
